third_party/blink/web_tests/platform/generic/http/tests/websocket/invalid-encode-length-expected.txt

Test whether WebSocket aborts the connection when it receives an unexpected encoding length frame.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Case 126_125: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Case 126_0: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Case 127_65535: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Case 127_0: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
